+++ extensions/search-npm/src/components/PackagListItem.tsx
@@ -97,6 +97,18 @@ export const PackageListItem = ({
         onOpen={handleAddToHistory}
       />
     ),
+    npmxPackagePage: (
+      <Action.OpenInBrowser
+        url={`https://npmx.dev/package/${pkg.name}`}
+        // eslint-disable-next-line @raycast/prefer-title-case
+        title="Open npmx Package Page"
+        icon={{
+          source: "npmx.png",
+        }}
+        key="npmxPackagePage"
+        onOpen={handleAddToHistory}
+      />
+    ),
   };
 
   const accessories: List.Item.Accessory[] = [